backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic<2.0.0,>=1.9.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
vecs==0.4.2
psycopg2==2.9.9
supabase>=1.0.3
httpx>=0.23.0,<0.24.0
numpy==1.26.4